+++ workoutpact/ShakeViewContorller.swift
@@ -39,6 +39,10 @@ class ShakeViewController: UIViewController {
             return
         }
 
+        if self.presentedViewController != nil {
+            return
+        }
+
         let alert = UIAlertController(title: "Share workout", message: "Post your completed workout to Twitter?", preferredStyle: UIAlertControllerStyle.Alert)
         alert.addAction(UIAlertAction(title: "Cancel", style: UIAlertActionStyle.Cancel, handler: nil))
         alert.addAction(UIAlertAction(title: "Share", style: UIAlertActionStyle.Default, handler: { (action) -> Void in